split out replicate tests from basic
